Switched to using a module-specific logger so code calling this module can have better control over logs.
This module uses Python's logging module for it's status messages (which is a good thing). However,
the way it was implemented totally ignored the logging needs of other code which call this module.
Because it used the 'root' logger and called BasicConfig() directly, it wasn't possible for other
code to control logging.

This version is functionally identical, except that it uses a module-specific "certstream" logger,
which can coexist peacefully with logging from other modules or the main program.